Register for COM Interop and New Guid.
1 In Solution Explorer, right-click the project name that is in bold type, and then click Properties.
Click the Build/Compile tab.
Scroll to the bottom of the Build/Compile window and select Register for COM Interop.
Note   "Register for COM Interop" causes the assembly to be automatically registered by the regasm.exe utility, when the project is compiled. So, you do not need to run regasm.exe manually from the command prompt.
From the File menu, click Save All.
Close the Compile/Build window.

2 used guid generator to generate to 2 guid and set them as  guid's of "EEPM-EEGateway class and EEPM-EECCValidator"